refactor: move tests to separated files
